fix(playlist): automate total playlist to create
Previously, the old total playlist to be created was hardcoded to 25.

This was because of mis-interpretation of splitVector function in Utils.

Utils Split Vector Function has been updated documentation and now automatically finds total number of playlist to create.

This is done via by TOTAL_SIZE / 50, because of hard limit on temporary playlist of YouTube [2023-04-08]

BREAKING CHANGE: Removes hardcoded function use and is better.