remove content from archive fill
The archive file that is provided should be empty and should be in fact be generated by the program software itself.

To achieve the filling of archive from the user end only it is better to use an empty file for the same.

The original custom files will be used to keep my original data added to gitignore.

Signed-off-by: Lakshya <manstein.felix@gmail.com>