give a huge update without track for files
Conclusive update to a lot of files that allows modularization.

This way of providing a single git update to a repo is bad as searching different functionality that were updated is difficult.

Also this brings a major problem in reverting to a version, merging and maintaing.

Also the title of git update must always mention the different parts of architecture that were updated to and what was modified, added or deleted.

Please take care.

Signed-off-by: Lakshya <manstein.felix@gmail.com>